requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.8.0=py37_0
astor=0.8.0=py37_0
blas=1.0=mkl
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2019.8.28=0
certifi=2019.9.11=py37_0
chardet=3.0.4=pypi_0
cycler=0.10.0=pypi_0
facenet=1.0.5=pypi_0
gast=0.3.2=py_0
google-pasta=0.1.7=py_0
grpcio=1.16.1=py37hf8bcb03_1
h5py=2.9.0=py37h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
idna=2.8=pypi_0
intel-openmp=2019.4=243
joblib=0.14.0=pypi_0
keras=2.3.0=pypi_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libprotobuf=3.9.2=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.1.1=py37_0
matplotlib=3.1.1=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.14=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mtcnn=0.0.9=pypi_0
ncurses=6.1=he6710b0_1
numpy=1.17.2=py37haad9e8e_0
numpy-base=1.17.2=py37hde5b4d6_0
opencv-python=4.1.1.26=pypi_0
openssl=1.1.1d=h7b6447c_2
pillow=6.2.0=pypi_0
pip=19.2.3=py37_0
protobuf=3.9.2=py37he6710b0_0
psutil=5.6.3=pypi_0
pyparsing=2.4.2=pypi_0
python=3.7.4=h265db76_1
python-dateutil=2.8.0=pypi_0
pyyaml=5.1.2=pypi_0
readline=7.0=h7b6447c_5
requests=2.22.0=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.3.1=py37h7c811a0_0
setuptools=41.2.0=py37_0
six=1.12.0=py37_0
sqlite=3.30.0=h7b6447c_0
tensorboard=1.14.0=py37hf484d3e_0
tensorflow=1.14.0=mkl_py37h45c423b_0
tensorflow-base=1.14.0=mkl_py37h7ce6ba3_0
tensorflow-estimator=1.14.0=py_0
termcolor=1.1.0=py37_1
tk=8.6.8=hbc83047_0
urllib3=1.25.6=pypi_0
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
